MAYA-109948 - Crash saving USD without opening Layer Editor
* Moved the MayaQtUtils init to the plugin.
